chore(node): move builtin-plugin related files to `builtin-plugin` folder
