+++ packages/rolldown/src/builtin-plugin/alias-plugin.ts
@@ -1,0 +1,22 @@
+import { BindingBuiltinPluginName } from '../binding'
+import { BuiltinPlugin } from './constructors'
+
+type AliasPluginAlias = {
+  find: string | RegExp
+  replacement: string
+}
+
+// A temp config type for giving better user experience
+type AliasPluginConfig = {
+  entries: AliasPluginAlias[]
+}
+
+class AliasPlugin extends BuiltinPlugin {
+  constructor(config?: AliasPluginConfig) {
+    super(BindingBuiltinPluginName.AliasPlugin, config)
+  }
+}
+
+export function aliasPlugin(config: AliasPluginConfig) {
+  return new AliasPlugin(config)
+}
