--- packages/rolldown/src/options/normalized-alias-plugin-config.ts
@@ -1,9 +1,0 @@
-type AliasPluginAlias = {
-  find: string | RegExp
-  replacement: string
-}
-
-// A temp config type for giving better user experience
-export type AliasPluginConfig = {
-  entries: AliasPluginAlias[]
-}
