+++ packages/rolldown/src/plugin/index.ts
@@ -20,7 +20,7 @@ import type { LogLevel } from '../log/logging'
 import type { RollupLog } from '../rollup'
 import type { MinimalPluginContext } from '../log/logger'
 import { InputOptions, OutputOptions } from '..'
-import { BuiltinPlugin } from './builtin-plugin'
+import { BuiltinPlugin } from '../builtin-plugin/constructors'
 import { ParallelPlugin } from './parallel-plugin'
 import type { DefinedHookNames } from '../constants/plugin'
 import { DEFINED_HOOK_NAMES } from '../constants/plugin'
