+++ packages/rolldown/src/utils/compose-js-plugins.ts
@@ -6,7 +6,7 @@ import {
 } from '../plugin'
 import { normalizeHook } from './normalize-hook'
 import { isNullish } from './misc'
-import { BuiltinPlugin } from '../plugin/builtin-plugin'
+import { BuiltinPlugin } from '../builtin-plugin/constructors'
 import { TupleToUnion } from 'type-fest'
 import * as R from 'remeda'
 import { PluginHookNames } from '../constants/plugin'
